+++ CHANGELOG.md
@@ -23,6 +23,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 
+ - Added logging of all emails in HTML format to `vagrant-magento/log/email`
  - Added host wrapper script for bin/magento command on guest
  - Added ability to modify guest config files (PHP, Apache etc) directly from host IDE
  - Added ability to choose if PhpStorm configs should be removed during project reinitialization